fix: Funcionamiento del servicio de clima.

+++ app/src/main/java/com/example/citymap/DTO/CoordinateDTO.kt
@@ -3,6 +3,6 @@ package com.example.citymap.DTO
 import com.google.gson.annotations.SerializedName
 
 data class CoordinateDTO(
-    @SerializedName("lat") var latitude: Float,
-    @SerializedName("lon") var longitude: Float
+    @SerializedName("lat") var latitude: Double,
+    @SerializedName("lon") var longitude: Double
 )
